dashboard/requirements.txt
gradio==6.2.0
hopsworks==4.2.0
pandas
plotly
python-dotenv
